Deleted all by Status

+++ src/main/java/com/chatapplicationspringBoot/Service/ChatService.java
@@ -8,8 +8,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * @author Talha Farooq
@@ -121,9 +123,13 @@ public class ChatService {
      */
     public ResponseEntity deleteChat(Long id) {
         try {
+            Optional<Chat> chatOptional = chatRepository.findById(id);
+            if(chatOptional.isPresent()){
             chatRepository.deleteById(id);
             logger.info("Deleteing Id");
             return new ResponseEntity(HttpStatus.OK);
+            }
+            else return new ResponseEntity("ID does not Exist", HttpStatus.NOT_FOUND);
         } catch (Exception e) {
             return new ResponseEntity("Cannot Find chat to delete", HttpStatus.NOT_FOUND);
         }
